requirements.txt
python-dotenv
deepgram-sdk
flask
groq